+++ src/db/getUserByUsername.ts
@@ -1,0 +1,11 @@
+import db from 'db';
+
+export function getUserByUsername(username: I.User['username']): I.User | null {
+   try {
+      return db.queryAll('users', {
+         query: { username },
+      })[0];
+   } catch {
+      return null;
+   }
+}
